backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.2
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.0
Faker==16.6.0
fastapi==0.89.1
frozenlist==1.3.3
greenlet==2.0.1
h11==0.14.0
idna==3.4
mailchimp3==3.0.17
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
passlib==1.7.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
yarl==1.8.2
pytz
